Add Range component for handling quantity range
